requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
dj-database-url==2.3.0
dj-email-url==1.0.6
Django==5.1.4
django-allauth==65.3.0
django-cache-url==3.4.5
django-crispy-forms==2.3
django-jalali-date==2.0.0
django-rosetta==0.10.1
environs==11.2.1
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
marshmallow==3.23.3
packaging==24.2
pillow==11.2.1
polib==1.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
